+++ src/ingest_validation_tools/upload.py
@@ -36,7 +36,6 @@ class ErrorDictException(Exception):
     def __init__(self, errors):
         message = f"Halting compilation of errors after detecting the following errors: {errors}."
         super().__init__(message)
-        # This returns only the error that caused the exception.
         labeled_errors = {}
         labeled_errors["Fatal Exception"] = errors
         self.errors = labeled_errors
@@ -57,6 +56,7 @@ class Upload:
         ignore_deprecation: bool = False,
         extra_parameters: Union[dict, None] = None,
         globus_token: str = "",
+        run_plugins: bool = False,
     ):
         self.directory_path = directory_path
         self.optional_fields = optional_fields
@@ -71,8 +71,8 @@ class Upload:
         self.effective_tsv_paths = {}
         self.extra_parameters = extra_parameters if extra_parameters else {}
         self.globus_token = globus_token
-
         # TODO: soft assay type endpoint calls happen here to weed out any bad multi-assay uploads
+        self.run_plugins = run_plugins
 
         try:
             unsorted_effective_tsv_paths = {
@@ -200,6 +200,11 @@ class Upload:
             reference_errors = self._get_reference_errors()
             if reference_errors:
                 errors["Reference Errors"] = reference_errors
+
+            if errors and not self.run_plugins:
+                raise ErrorDictException(
+                    "Skipping plugins validation: errors in upload metadata or dir structure."
+                )
 
             plugin_errors = self._get_plugin_errors(**kwargs)
             if plugin_errors: