+++ tests-manual/test-dataset-examples-cedar.sh
@@ -12,7 +12,7 @@ for SUITE in examples/dataset-examples examples/dataset-iec-examples; do
             OPTS="--dataset_ignore_globs 'metadata.tsv' --upload_ignore_globs '*'"
             ;;
         examples/dataset-examples)
-            OPTS="--dataset_ignore_globs 'ignore-*.tsv' '.*' --globus_token ${GLOBUS_TOKEN} --upload_ignore_globs 'drv_ignore_*' --output as_md"
+            OPTS="--dataset_ignore_globs 'ignore-*.tsv' '.*' --run_plugins --globus_token ${GLOBUS_TOKEN} --upload_ignore_globs 'drv_ignore_*' --output as_md"
             ;;
         *)
             die "Unexpected ${SUITE}"